Mostly cosmetic use of uintptr_t to hold addresses instead of uint32_t

+++ arch/arm/include/armv7-a/irq.h
@@ -250,7 +250,7 @@ struct xcptcontext
    */
 
 #if 0 /* Not yet implemented */
-  FAR uint32_t *stack[ARCH_STACK_NSECTS];
+  FAR uintptr_t *stack[ARCH_STACK_NSECTS];
 #endif
 #endif
 };
